# services/errotel/internal/oas/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn11AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn3AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn10AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn4AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn12AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn6AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
	rn7AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"GET": "Authorization",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/v1/"

			if l := len("/api/v1/"); len(elem) >= l && elem[0:l] == "/api/v1/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "alerts/prepare"

				if l := len("alerts/prepare"); len(elem) >= l && elem[0:l] == "alerts/prepare" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handlePrepareAlertRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn11AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'c': // Prefix: "capabilities"

				if l := len("capabilities"); len(elem) >= l && elem[0:l] == "capabilities" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetCapabilitiesRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn1AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'f': // Prefix: "facets"

				if l := len("facets"); len(elem) >= l && elem[0:l] == "facets" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleGetFacetsRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'i': // Prefix: "investigate"

				if l := len("investigate"); len(elem) >= l && elem[0:l] == "investigate" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleInvestigateRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn10AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'o': // Prefix: "occurrences/"

				if l := len("occurrences/"); len(elem) >= l && elem[0:l] == "occurrences/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'h': // Prefix: "histogram"
					origElem := elem
					if l := len("histogram"); len(elem) >= l && elem[0:l] == "histogram" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleGetHistogramRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn4AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				case 's': // Prefix: "search"
					origElem := elem
					if l := len("search"); len(elem) >= l && elem[0:l] == "search" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSearchOccurrencesRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn12AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				}
				// Param: "ref"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleGetOccurrenceRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn6AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/related"

					if l := len("/related"); len(elem) >= l && elem[0:l] == "/related" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleGetRelatedLogsRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn7AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

				}

			case 't': // Prefix: "traces/"

				if l := len("traces/"); len(elem) >= l && elem[0:l] == "traces/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "traceId"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetTraceRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn9AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/v1/"

			if l := len("/api/v1/"); len(elem) >= l && elem[0:l] == "/api/v1/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "alerts/prepare"

				if l := len("alerts/prepare"); len(elem) >= l && elem[0:l] == "alerts/prepare" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = PrepareAlertOperation
						r.summary = ""
						r.operationID = "prepareAlert"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/alerts/prepare"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'c': // Prefix: "capabilities"

				if l := len("capabilities"); len(elem) >= l && elem[0:l] == "capabilities" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetCapabilitiesOperation
						r.summary = ""
						r.operationID = "getCapabilities"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/capabilities"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'f': // Prefix: "facets"

				if l := len("facets"); len(elem) >= l && elem[0:l] == "facets" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = GetFacetsOperation
						r.summary = ""
						r.operationID = "getFacets"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/facets"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'i': // Prefix: "investigate"

				if l := len("investigate"); len(elem) >= l && elem[0:l] == "investigate" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = InvestigateOperation
						r.summary = ""
						r.operationID = "investigate"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/investigate"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'o': // Prefix: "occurrences/"

				if l := len("occurrences/"); len(elem) >= l && elem[0:l] == "occurrences/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'h': // Prefix: "histogram"
					origElem := elem
					if l := len("histogram"); len(elem) >= l && elem[0:l] == "histogram" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = GetHistogramOperation
							r.summary = ""
							r.operationID = "getHistogram"
							r.operationGroup = ""
							r.pathPattern = "/api/v1/occurrences/histogram"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				case 's': // Prefix: "search"
					origElem := elem
					if l := len("search"); len(elem) >= l && elem[0:l] == "search" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = SearchOccurrencesOperation
							r.summary = ""
							r.operationID = "searchOccurrences"
							r.operationGroup = ""
							r.pathPattern = "/api/v1/occurrences/search"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "ref"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = GetOccurrenceOperation
						r.summary = ""
						r.operationID = "getOccurrence"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/occurrences/{ref}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/related"

					if l := len("/related"); len(elem) >= l && elem[0:l] == "/related" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = GetRelatedLogsOperation
							r.summary = ""
							r.operationID = "getRelatedLogs"
							r.operationGroup = ""
							r.pathPattern = "/api/v1/occurrences/{ref}/related"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 't': // Prefix: "traces/"

				if l := len("traces/"); len(elem) >= l && elem[0:l] == "traces/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "traceId"
				// Leaf parameter, slashes are prohibited
				idx := strings.IndexByte(elem, '/')
				if idx >= 0 {
					break
				}
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetTraceOperation
						r.summary = ""
						r.operationID = "getTrace"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/traces/{traceId}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}